+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fix issue with 'Enter' key in circulation rules editor that adds indent to previous line. Refs UICIRC-1042.
 * RulesEditor should use stripes.hasPerm. Refs UICIRC-1047.
 * Disable block template under reminder fees. Refs UICIRC-1046.
+* UI tests replacement with RTL/Jest for `CirculationRules.js`. Refs UICIRC-958.
 
 ## [9.0.3](https://github.com/folio-org/ui-circulation/tree/v9.0.3) (2023-11-09)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v9.0.2...v9.0.3)